# docs/api/openapi.yaml
components:
  schemas:
...
    AccountTypeRequest:
      properties:
        account_type:
          pattern: ^(individual|legal)$
          title: Account Type
          type: string
        actual_address:
          anyOf:
          - type: string
          - type: 'null'
          title: Actual Address
        bank_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Bank Name
        bik:
          anyOf:
          - type: string
          - type: 'null'
          title: Bik
        checking_account:
          anyOf:
          - type: string
          - type: 'null'
          title: Checking Account
        company_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Company Name
        corr_account:
          anyOf:
          - type: string
          - type: 'null'
          title: Corr Account
        director_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Director Name
        docs_email:
          anyOf:
          - type: string
          - type: 'null'
          title: Docs Email
        full_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Full Name
        inn:
          anyOf:
          - type: string
          - type: 'null'
          title: Inn
        kpp:
          anyOf:
          - type: string
          - type: 'null'
          title: Kpp
        legal_address:
          anyOf:
          - type: string
          - type: 'null'
          title: Legal Address
        ogrn:
          anyOf:
          - type: string
          - type: 'null'
          title: Ogrn
      required:
      - account_type
      title: AccountTypeRequest
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    TopupRequest:
      properties:
        amount:
          default: 1000
          maximum: 500000.0
          minimum: 100.0
          title: Amount
          type: integer
        customer_name:
          anyOf:
          - maxLength: 255
            type: string
          - type: 'null'
          title: Customer Name
        payment_method:
          default: redirect
          pattern: ^(redirect|sbp_qr|card|sbp)$
          title: Payment Method
          type: string
      title: TopupRequest
      type: object
...
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http
info:
  description: API платформы 3D-моделей для маркетплейсов
  title: KWork Mob API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/user/account-type:
    post:
      operationId: account_type_api_v1_user_account_type_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountTypeRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Account Type
      tags:
      - Пользователь
  /api/v1/user/balance:
    get:
      operationId: get_balance_api_v1_user_balance_get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: Get Balance
      tags:
      - Пользователь
  /api/v1/user/balance-filters:
    get:
      description: Saved personal transaction filters §20.3.4.
      operationId: get_balance_filters_api_v1_user_balance_filters_get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: Get Balance Filters
      tags:
      - Пользователь
    put:
      description: Persist personal transaction filters §20.3.4.
      operationId: put_balance_filters_api_v1_user_balance_filters_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BalanceFiltersBody'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Put Balance Filters
      tags:
      - Пользователь
  /api/v1/user/balance/payment/{payment_id}:
    get:
      description: Polling статуса пополнения (СБП QR / карта) §20.3.3.
      operationId: topup_payment_status_api_v1_user_balance_payment__payment_id__get
      parameters:
      - in: path
        name: payment_id
        required: true
        schema:
          title: Payment Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Topup Payment Status
      tags:
      - Пользователь
  /api/v1/user/balance/topup:
    post:
      description: 'Пополнение: карта (redirect) или СБП QR + фискальный чек (§8.12
        / §8.6.4).'
      operationId: topup_balance_api_v1_user_balance_topup_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TopupRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Topup Balance
      tags:
      - Пользователь
  /api/v1/user/devices:
    delete:
      operationId: unregister_device_api_v1_user_devices_delete
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceTokenRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Unregister Device
      tags:
      - Пользователь
    post:
      description: Регистрация FCM/APNs токена (§3.4.3).
      operationId: register_device_api_v1_user_devices_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceTokenRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Register Device
      tags:
      - Пользователь
  /api/v1/user/draft-backups:
    get:
      description: Список облачных черновиков TTL 7 дней §3.3.2.
      operationId: list_draft_backups_api_v1_user_draft_backups_get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: List Draft Backups
      tags:
      - Пользователь
  /api/v1/user/draft-backups/prepare:
    post:
      description: Presigned upload ZIP черновика.
      operationId: prepare_draft_backup_api_v1_user_draft_backups_prepare_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DraftBackupPrepareBody'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Prepare Draft Backup
      tags:
      - Пользователь
  /api/v1/user/draft-backups/{model_uuid}/restore:
    get:
      description: Presigned download для восстановления черновика §3.3.2.
      operationId: restore_draft_backup_api_v1_user_draft_backups__model_uuid__restore_get
      parameters:
      - in: path
        name: model_uuid
        required: true
        schema:
          title: Model Uuid
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Restore Draft Backup
      tags:
      - Пользователь
  /api/v1/user/me:
    get:
      description: Текущий пользователь (для seller и staff).
      operationId: get_me_api_v1_user_me_get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: Get Me
      tags:
      - Пользователь
    patch:
      description: Обновление профиля (ФИО, телефон, маркетинг).
      operationId: update_me_api_v1_user_me_patch
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: true
              title: Payload
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Update Me
      tags:
      - Пользователь
  /api/v1/user/me/delete-request:
    post:
      description: 'Право на забвение: заявка, SLA 30 дней (§2.8.3).'
      operationId: request_account_deletion_api_v1_user_me_delete_request_post
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: Request Account Deletion
      tags:
      - Пользователь
  /api/v1/user/models:
    get:
      operationId: list_user_models_api_v1_user_models_get
      parameters:
      - in: query
        name: company_id
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Company Id
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: List User Models
      tags:
      - Пользователь
  /api/v1/user/notifications:
    delete:
      operationId: notifications_clear_api_v1_user_notifications_delete
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: Notifications Clear
      tags:
      - Пользователь
    get:
      description: Inbox уведомлений §19.16.
      operationId: list_notifications_api_v1_user_notifications_get
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          maximum: 100
          minimum: 1
          title: Limit
          type: integer
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          title: Offset
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: List Notifications
      tags:
      - Пользователь
  /api/v1/user/notifications/read-all:
    post:
      operationId: notifications_read_all_api_v1_user_notifications_read_all_post
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: Notifications Read All
      tags:
      - Пользователь
  /api/v1/user/notifications/{notification_id}/read:
    post:
      operationId: notification_mark_read_api_v1_user_notifications__notification_id__read_post
      parameters:
      - in: path
        name: notification_id
        required: true
        schema:
          title: Notification Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Notification Mark Read
      tags:
      - Пользователь
...